Autorizacion Familias profesionales e idiomas

+++ app/Http/Controllers/API/FamiliaProfesionalController.php
@@ -12,6 +12,11 @@ class FamiliaProfesionalController extends Controller
 {
     public $modelclass = FamiliaProfesional::class;
 
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show']);
+        $this->authorizeResource(FamiliaProfesional::class, 'familiaProfesional');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -29,6 +34,7 @@ class FamiliaProfesionalController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', FamiliaProfesional::class);
         $familiaProfesional = json_decode($request->getContent(), true);
 
         $familiaProfesional = FamiliaProfesional::create($familiaProfesional);
@@ -49,6 +55,7 @@ class FamiliaProfesionalController extends Controller
      */
     public function update(Request $request, FamiliaProfesional $familiaProfesional)
     {
+        $this->authorize('update', $familiaProfesional);
         $familiaProfesionalData = json_decode($request->getContent(), true);
         $familiaProfesional->update($familiaProfesionalData);
 
@@ -60,6 +67,7 @@ class FamiliaProfesionalController extends Controller
      */
     public function destroy(FamiliaProfesional $familiaProfesional)
     {
+        $this->authorize('delete', $familiaProfesional);
         $familiaProfesional->delete();
     }
 }
