+++ app/Http/Controllers/API/IdiomaController.php
@@ -11,6 +11,12 @@ use Illuminate\Http\Request;
 class IdiomaController extends Controller
 {
     public $modelclass = Idioma::class;
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show']);
+        $this->authorizeResource(Idioma::class, 'idioma');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -28,6 +34,7 @@ class IdiomaController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Idioma::class);
         $idioma = json_decode($request->getContent(), true);
 
         $idioma = Idioma::create($idioma);
@@ -48,6 +55,7 @@ class IdiomaController extends Controller
      */
     public function update(Request $request, Idioma $idioma)
     {
+        $this->authorize('update', $idioma);
         $idiomaData = json_decode($request->getContent(), true);
 
         $idioma->update($idiomaData);
@@ -60,6 +68,7 @@ class IdiomaController extends Controller
      */
     public function destroy(Idioma $idioma)
     {
+        $this->authorize('delete', $idioma);
         $idioma->delete();
     }
 }
